study the idea of greedy quasiswapping
We always pick the first row & column among the rows and columns
that have >1 elements. This does not guarantee that 2x2 submatrix
has such a element (except at the end where it really matters), but
increases the probability. This adds a lot of ugly bookkeeping.

Syntactically correct, but no API yet and therefore untested.